+++ src/main/java/net/zaltren/rcna/common/CommonProxy.java
@@ -2,7 +2,13 @@ package net.zaltren.rcna.common;
 
 public class CommonProxy {
 
+    public void preInit() {
+    }
+
     public void init() {
+    }
+
+    public void postInit() {
     }
 
 }